Add Makefile caching helpers for lazy, expensive evaluations
Makefiles natively offer two variable expansion modes: immediate and
deferred.  When expanding variables that require invocations of external
programs (such as `llvm-config`) immediate expansion is almost always
preferred, as it will run the external command once, exactly when the
makefile variable is defined.  Deferred mode, on the other hand, will
expand the variable every time it is used, running the external program
again and again.

When the external program is expensive, this cost can slow down the
build significantly, however when the external program requires some
setup (for instance, when it itself is downloaded through other rules in
the Makefile) it cannot always be immediately expanded.  To address
this, we build a caching layer that allows for deferred expansion, but
once it has been expanded once, the variable is replaced with the result
of running the command, and further hits to the same variable will
return the cached value.  (With the slight caveat that an empty result
will cause the external command to be run again in the future).